fix(web): move space policy value mapping into transformations.ts
fields.ts should hold only the form vocabulary (types, constants,
defaults). Move toFormValues/buildPayload (renamed from
buildInitialValues/buildActionsPayload) into transformations.ts, matching
the form code-organization convention and the partition-form layout.

Make that rule explicit in form/conventions.md: value mapping lives in
transformations.ts, not fields.ts.
